+++ jannovar-vardbs/src/main/java/de/charite/compbio/jannovar/vardbs/facade/DBVariantContextAnnotatorFactory.java
@@ -3,9 +3,9 @@ package de.charite.compbio.jannovar.vardbs.facade;
 import de.charite.compbio.jannovar.vardbs.base.DBAnnotationOptions;
 import de.charite.compbio.jannovar.vardbs.base.JannovarVarDBException;
 import de.charite.compbio.jannovar.vardbs.clinvar.ClinVarAnnotationDriver;
+import de.charite.compbio.jannovar.vardbs.cosmic.CosmicAnnotationDriver;
 import de.charite.compbio.jannovar.vardbs.dbsnp.DBSNPAnnotationDriver;
 import de.charite.compbio.jannovar.vardbs.exac.ExacAnnotationDriver;
-import de.charite.compbio.jannovar.vardbs.uk10k.UK10KAnnotationDriver;
 import htsjdk.variant.variantcontext.VariantContext;
 
 /**
@@ -66,7 +66,7 @@ public class DBVariantContextAnnotatorFactory {
 	 */
 	public DBVariantContextAnnotator constructUK10K(String vcfDBPath, String fastaRefPath, DBAnnotationOptions options)
 			throws JannovarVarDBException {
-		return new DBVariantContextAnnotator(new UK10KAnnotationDriver(vcfDBPath, fastaRefPath, options), options);
+		return new DBVariantContextAnnotator(new CosmicAnnotationDriver(vcfDBPath, fastaRefPath, options), options);
 	}
 
 	/**